@@ -8,8 +8,12 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "axios": "^1.4.0",
+    "bootstrap": "^4.0.0",
     "core-js": "^3.8.3",
-    "vue": "^2.6.14"
+    "vue": "^3.2.0",
+    "vue-resource": "^1.5.3",
+    "vue-router": "^4.2.4"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
